requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Werkzeug
PyJWT
numpy
Pillow
python-dotenv
gunicorn
psycopg2-binary